Problem: the project structure is not decoupled.
Currently, the different parts of PumpkinDB all reside in
a single crate, enforcing a tight coupling between all different parts.

Solution: Restructure the code into several sub-crates, using
the Cargo workspace feature:

* pumpkinscript: Parser for the script language
* pumpkindb_engine: core database engine
* pumpkindb_server: mio based server on top of the engine
* pumpkindb_term: terminal client

Integration tests will reside in sub-crates inside the new /tests
directory.
The doctests were already moved there.

This also updates travis + AppVeyor config files and the README
with new build instructions.
