+++ CONTRIBUTORS
@@ -1,3 +1,4 @@
 Guillaume Gauvrit <guillaume@gauvr.it>
 Ómar Yasin <omarkj@gmail.com>
 Yurii Rashkovskii <yrashk@gmail.com>
+Christoph Herzog <chris@theduke.at>
