+++ appveyor.yml
@@ -21,12 +21,10 @@ install:
 build: false
 
 test_script:
-  - cargo build --verbose %cargoflags%
-  - cargo build --features="experimental" --verbose
-  - cargo test -- --nocapture
-  - cargo test --features="experimental" -- --nocapture
-  - cargo test --features="scoped_dictionary" -- --nocapture
-  - cargo test --features="static_module_dispatch" -- --nocapture
-  - cargo run --bin pumpkindb-doctests
-  - cargo run --bin pumpkindb-doctests --features="experimental"
-  - cargo run --bin pumpkindb-doctests --features="static_module_dispatch"
+  - cargo build --all --verbose %cargoflags%
+  - cargo build --all --features="experimental" --verbose
+  - cargo test --all -- --nocapture
+  - cargo test --all --features="experimental" -- --nocapture
+  - cargo test --all --features="scoped_dictionary" -- --nocapture
+  - cargo test --all --features="static_module_dispatch" -- --nocapture
+  - target\debug\pumpkindb-doctests
