+++ pumpkindb_engine/Cargo.toml
@@ -1,0 +1,50 @@
+[package]
+name = "pumpkindb_engine"
+version = "0.2.0"
+license = "MPL-2.0"
+repository = "https://github.com/PumpkinDB/PumpkinDB"
+homepage = "http://pumpkindb.org"
+keywords = [ "pumpkindb", "database" ]
+categories = [ "database" ]
+
+authors = ["Guillaume Gauvrit <guillaume@gauvr.it>",
+           "Ómar Yasin <omarkj@gmail.com>",
+           "Yurii Rashkovskii <yrashk@gmail.com>"]
+
+[dependencies]
+nom = "^2.1"
+snowflake = { git = "https://github.com/yrashk/snowflake.git", branch="pub-fields" }
+lmdb-zero = "0.4.0"
+config = "0.3.1"
+lazy_static = "0.2.2"
+num-bigint = "0.1.35"
+num-traits = "0.1.36"
+libc = "0.2.20"
+hybrid-clocks = "0.3.2"
+byteorder = "1.0.0"
+uuid = { version = "0.4.0", features = ["v4"] }
+glob = "0.2.11"
+regex = "0.2.1"
+crossbeam = "0.2.10"
+tempdir = "0.3.5"
+rust-crypto = "^0.2"
+log = "0.3.6"
+log4rs = { version = "0.6.1", features = ["toml_format"] }
+serde_json = "0.9.8"
+clap = "2.20.5"
+num_cpus = "1.3.0"
+rand = "0.3.15"
+memmap = "0.5.2"
+
+pumpkinscript = { version = "0.2", path = "../pumpkinscript" }
+
+[dev-dependencies]
+quickcheck = "0.4.1"
+quickcheck_macros = "0.4.1"
+matches = "0.1.4"
+
+[features]
+default = ["scoped_dictionary"]
+experimental = []
+scoped_dictionary = []
+static_module_dispatch = []
