+++ pumpkindb_engine/src/script/macros.rs
@@ -4,36 +4,6 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-macro_rules! write_size_into_slice {
-    ($size:expr, $slice: expr) => {
-     match $size {
-        0...120 => {
-            $slice[0] = $size as u8;
-            1
-        }
-        121...255 => {
-            $slice[0] = 121u8;
-            $slice[1] = $size as u8;
-            2
-        }
-        256...65535 => {
-            $slice[0] = 122u8;
-            $slice[1] = ($size >> 8) as u8;
-            $slice[2] = $size as u8;
-            3
-        }
-        65536...4294967296 => {
-            $slice[0] = 123u8;
-            $slice[1] = ($size >> 24) as u8;
-            $slice[2] = ($size >> 16) as u8;
-            $slice[3] = ($size >> 8) as u8;
-            $slice[4] = $size as u8;
-            5
-        }
-        _ => unreachable!(),
-    }
-    };
-}
 
 macro_rules! handle_error {
     ($env: expr, $err: expr) => {
@@ -194,20 +164,6 @@ macro_rules! error_unknown_instruction {
             ERROR_UNKNOWN_INSTRUCTION
         )
     } }
-}
-
-macro_rules! write_size_header {
-    ($bytes: expr, $vec: expr) => {{
-        write_size!($bytes.len(), $vec);
-    }};
-}
-
-macro_rules! write_size {
-    ($size: expr, $vec: expr) => {{
-        let mut header = vec![0;offset_by_size($size)];
-        write_size_into_slice!($size, header.as_mut_slice());
-        $vec.append(&mut header);
-    }};
 }
 
 macro_rules! alloc_slice {
