+++ pumpkindb_engine/src/script/mod_binaries.rs
