+++ pumpkindb_server/Cargo.toml
@@ -1,0 +1,35 @@
+[package]
+name = "pumpkindb_server"
+version = "0.2.0"
+license = "MPL-2.0"
+repository = "https://github.com/PumpkinDB/PumpkinDB"
+homepage = "http://pumpkindb.org"
+keywords = [ "pumpkindb", "database" ]
+categories = [ "database" ]
+
+authors = ["Guillaume Gauvrit <guillaume@gauvr.it>",
+           "Ómar Yasin <omarkj@gmail.com>",
+           "Yurii Rashkovskii <yrashk@gmail.com>"]
+
+[[bin]]
+doc = false
+name = "pumpkindb"
+
+[dependencies]
+nom = "^2.1"
+mio = "0.6.4"
+byteorder = "1.0.0"
+memmap = "0.5.2"
+slab = "0.3.0"
+num-bigint = "0.1.35"
+num-traits = "0.1.36"
+rand = "0.3.15"
+num_cpus = "1.3.0"
+log = "0.3.6"
+log4rs = { version = "0.6.1", features = ["toml_format"] }
+config = "0.3.1"
+lazy_static = "0.2.2"
+lmdb-zero = "0.4.0"
+
+pumpkinscript = { version = "0.2", path = "../pumpkinscript" }
+pumpkindb_engine = { version = "0.2", path = "../pumpkindb_engine" }
