+++ pumpkindb_term/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "pumpkindb_term"
+version = "0.2.0"
+license = "MPL-2.0"
+repository = "https://github.com/PumpkinDB/PumpkinDB"
+homepage = "http://pumpkindb.org"
+keywords = [ "pumpkindb", "database" ]
+
+authors = ["Guillaume Gauvrit <guillaume@gauvr.it>",
+           "Ómar Yasin <omarkj@gmail.com>",
+           "Yurii Rashkovskii <yrashk@gmail.com>"]
+
+[[bin]]
+doc = false
+name = "pumpkindb-term"
+
+[dependencies]
+config = "0.3.1"
+nom = "^2.1"
+rustyline = "1.0.0"
+ansi_term = "0.9.0"
+uuid = { version = "0.4.0", features = ["v4"] }
+clap = "2.20.5"
+byteorder = "1.0.0"
+
+pumpkinscript = { version = "0.2", path = "../pumpkinscript" }
+pumpkindb_engine = { version = "0.2", path = "../pumpkindb_engine" }
