+++ tests/doctests/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "doctests"
+version = "0.1.0"
+
+[[bin]]
+doc = false
+name = "pumpkindb-doctests"
+
+[dependencies]
+glob = "0.2.11"
+regex = "0.2.1"
+crossbeam = "0.2.10"
+tempdir = "0.3.5"
+lmdb-zero = "0.4.0"
+nom = "^2.1"
+
+pumpkinscript = { version = "0.2", path = "../../pumpkinscript" }
+pumpkindb_engine = { version = "0.2", path = "../../pumpkindb_engine" }
